feat: refactor codebase and improve command handling
This commit refactors the codebase and improves command handling. Changes include:

- Renamed several files to follow a new naming convention
- Moved command handling logic to a separate function for better readability and maintainability
- Updated `getFiles` function in utils to support optional recursive file loading
- Updated client structure to store categories of commands
- Improved logging for registered commands
- Updated import paths for several dependencies
