+++ src/structures/Embed.ts
